+++ src/Sulu/Bundle/ContentBundle/Resources/public/js/components/single-internal-link/main.js
@@ -205,7 +205,7 @@ define([], function() {
         },
 
         getSelectUrl = function(url, data) {
-            return url.replace('{id=dataSource&}', (!!data ? 'id=' + data + '&' : ''));
+            return url.replace('{id=uuid&}', (!!data ? 'id=' + data + '&' : ''));
         },
 
         getSingleUrl = function(url, data) {